SaveChangeAsync Interceptor
Base Entity e yeni prop eklendi ve migrationlar baştan basıldı hatalar düzeltildi.SaveChangeAsync için Interceptor  yapısı kuruldu.Ne zaman herangi bir veri eklendiğinde veya güncellendiğinde savechanges tetikelndiğinde ilk önce override tetiklenicek gelen dataları yakalıcak statelere göre dateTime proplarını düzenlicek .böylelikle sürekli manual olarak date time nesnesi yaratmamıza gerek olmayacak.

+++ Core/ETicaretAPI.Domain/Entities/Product.cs
@@ -9,7 +9,7 @@ namespace ETicaretAPI.Domain.Entities
 
 		public int Stock { get; set; }
 
-		public long Price { get; set; }
+		public float Price { get; set; }
 
 		public ICollection<Order> Orders { get; set; }
 	}
